make countup a client component
