+++ src/app/api/oev-data/route.ts
@@ -10,17 +10,17 @@ const QUERY_ID = 3373179;
 export async function GET() {
   try {
     const executionRes = await Dune.refresh(QUERY_ID);
-    console.log(executionRes)
+    console.log(executionRes);
 
     if (!executionRes.result) {
       throw new Error("Failed to execute query");
     }
 
-    const oevLost = (executionRes.result.rows[0] as {'max_potential_revenue_usd': number}).max_potential_revenue_usd
+    const oevLost = (executionRes.result.rows[0] as { max_potential_revenue_usd: number }).max_potential_revenue_usd;
     // cache successful response
     return NextResponse.json({
       status: 200,
-      body: { oevLost }, // find where value lives
+      oevLost, // find where value lives
     });
   } catch (e) {
     console.error(e);
@@ -31,7 +31,3 @@ export async function GET() {
     });
   }
 }
-
-export type OevDataResponse = {
-  oevLost: number;
-};
