+++ src/app/api/oev-data/utils.ts
@@ -1,0 +1,3 @@
+export type OevDataResponse = {
+  oevLost: number;
+};
